refactor gallery-masonry block
